+++ package.json
@@ -41,15 +41,14 @@
     "test-server": "mocha --recursive test/server",
     "test-server-coverage": "nyc --reporter=lcov --reporter=text mocha --recursive test/server",
     "eslint": "eslint .",
-    "tslint": "tslint ./js/react/**/*.{ts,tsx}",
+    "tslint": "tslint ts/**/*.{ts,tsx}",
     "transpile": "tsc",
-    "clean-transpile": "rimraf js/react/**/*.js js/react/*.js",
+    "clean-transpile": "rimraf ts/**/*.js ts/*.js",
     "open-coverage": "open coverage/lcov-report/index.html",
     "styleguide": "styleguidist server"
   },
   "dependencies": {
     "@sindresorhus/is": "^0.8.0",
-    "@types/qs": "^6.5.1",
     "archiver": "^2.1.1",
     "blob-util": "^1.3.0",
     "blueimp-canvas-to-blob": "^3.14.0",
@@ -92,6 +91,7 @@
     "websocket": "^1.0.25"
   },
   "devDependencies": {
+    "@types/qs": "^6.5.1",
     "@types/react": "^16.3.1",
     "@types/react-dom": "^16.0.4",
     "arraybuffer-loader": "^1.0.3",
