+++ test/styleguide/legacy_bridge.js
@@ -1,5 +1,6 @@
+/* global window: false */
 
-// Because we aren't hosting the styleguide in Electron, we can't rely on preload.js
+// Because we aren't hosting the Style Guide in Electron, we can't rely on preload.js
 //   to set things up for us. This gives us the minimum bar shims for everything it
 //   provdes.
 //
@@ -17,14 +18,14 @@ window.Signal.Migrations = {
   getPlaceholderMigrations: () => {},
 };
 
-window.Signal.React = window.Signal.React = {};
+window.Signal.React = {};
 
 window.EmojiConvertor = function EmojiConvertor() {};
 window.EmojiConvertor.prototype.init_colons = () => {};
 window.EmojiConvertor.prototype.signalReplace = html => html;
 window.EmojiConvertor.prototype.replace_unified = string => string;
 window.EmojiConvertor.prototype.img_sets = {
-  apple: {}
+  apple: {},
 };
 
 window.i18n = () => '';
