+++ test/styleguide/legacy_templates.js
@@ -1,20 +1,27 @@
+/* global window: false */
 
 // Taken from background.html.
-// Templates are here solely to support the Backbone views rendered in the styleguide.
+// Templates are here solely to support the Backbone views rendered in the Style Guide.
+
+// Note: Any change here must be reflected in background.html to be reflected in the app
+//   and test/index.html to be reflected in the unit tests.
 
 window.Whisper.View.Templates = {
   hasRetry: `
-    {{ messageNotSent }}
-    <span href='#' class='retry'>{{ resend }}</span>
+    {{ messageNotSent }} <span href='#' class='retry'>{{ resend }}</span>
   `,
   'some-failed': `
     {{ someFailed }}
   `,
   keychange: `
-    <span class='content' dir='auto'><span class='shield icon'></span> {{ content }}</span>
+    <span class='content' dir='auto'>
+      <span class='shield icon'></span> {{ content }}
+    </span>
   `,
   'verified-change': `
-    <span class='content' dir='auto'><span class='{{ icon }} icon'></span> {{ content }}</span>
+    <span class='content' dir='auto'>
+      <span class='{{ icon }} icon'></span> {{ content }}
+    </span>
   `,
   message: `
     {{> avatar }}
@@ -41,5 +48,5 @@ window.Whisper.View.Templates = {
   `,
   expirationTimerUpdate: `
     <span class='content'><span class='icon clock'></span> {{ content }}</span>
-  `
+  `,
 };
